visuals/requirements_dev.txt
# Development environment pip requirements file
# Install via 'pip install -r requirements_dev.txt'

Flask
jinja2
simplejson
markdown
Flask-Markdown
pyyaml